tis: addressing potential int underflow
Under review it was determined that a malicious or erroneous TPM may
return a size equal to that of the header which will make it past
current guards but will result in an integer underflow. To address this
an additional guard is added after reading in the response header that
verifies the size field is larger than header before subtracting header
from the response size.

Signed-off-by: Daniel P. Smith <dpsmith@apertussolutions.com>