core: refactored hw interface
This commit refactors how the library interfaces with the hardware
interfaces. In particular it abstracts the interfaces into an
operations callback interface on struct tpm. This eliminates a lot
of the switches on tpm->intf to select the correct function for the
hardware interface. Additionally this refactoring enabled cleaning
up some areas of the code logic.

Signed-off-by: Daniel P. Smith <dpsmith@apertussolutions.com>